Initial working toast messages
- Functioning success messages.
- Move utils file to templates folder.
- Still to update failure message / explore greater modularity.
